# lib/api-zod/src/generated/types/campaignFilterOptions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CountryEntry } from "./countryEntry";

export interface CampaignFilterOptions {
  countries: CountryEntry[];
  categories: string[];
}